+++ src/NadekoBot/Modules/Music/_common/Impl/AudioFileCache/AudioFileCacheConfig.cs
@@ -4,7 +4,7 @@ namespace NadekoBot.Modules.Music;
 
 public sealed class AudioFileCacheConfig
 {
-    [Comment("DO NOT CHANGE")]
+    [Comment("DO NOT CHANGE THE VERSION MANUALLY")]
     public int Version { get; set; } = 1;
 
     [Comment("Maximum total cache size in gigabytes. Minimum 1. Default 10")]